+++ lib/src/utils.dart
@@ -102,11 +102,11 @@ List<double> multiplyMatrix5(List<double> a, List<double> b) {
   ];
 }
 
-ColorFilter toColorFilterMatrix(matrix) {
+ColorFilter toColorFilterMatrix(List<double> matrix) {
   return ColorFilter.matrix(matrix.sublist(0, 20));
 }
 
-Widget execFilterSample(matrix, child) {
+Widget execFilterSample(List<double> matrix, Widget child) {
   return ColorFiltered(colorFilter: toColorFilterMatrix(matrix), child: child);
 }
 
@@ -116,4 +116,87 @@ bool isNotNegative(double v) {
 
 bool isNotDefault(double v) {
   return v != 1.0 && v >= 0.0;
+}
+
+execShaderDirectSample(Color color, [Alignment end = Alignment.centerRight]) {
+  return (Rect bounds) {
+    return LinearGradient(
+      end: end,
+      colors: [color, color],
+      stops: const [0.0, 1.0]
+    ).createShader(bounds);
+  };
+}
+
+execShaderLinearSample(List<Color> colors, [Alignment end = Alignment.centerRight, List<double> stops = const [0.0, 1.0]]) {
+  return (Rect bounds) {
+    return LinearGradient(
+      end: end,
+      colors: colors,
+      stops: stops
+    ).createShader(bounds);
+  };
+}
+
+execShaderRadialSample(List<Color> colors, [List<double> stops = const [0.0, 1.0], radius = 0.8]) {
+  return (Rect bounds) {
+    return RadialGradient(
+      center: Alignment.center,
+      radius: radius,
+      colors: colors,
+      stops: stops
+      // tileMode: TileMode.mirror,
+    ).createShader(bounds);
+  };
+}
+
+/// Generates the configuration for applying CSSFilter effects, which is provided to `CSSFilter.apply` for use.
+/// Supports chain calls.
+/// 
+/// Example:
+/// 
+/// ```dart
+/// CSSFilter.apply(
+///   child: const Text('Hello World!'),
+///   value: CSSFilterMatrix().contrast(1.5).sepia(0.4)
+/// );
+/// ```
+class CSSFilterMatrix {
+  Map conf = {};
+  contrast([double value = 1.0]) {
+    conf['contrast'] = value;
+    return this;
+  }
+  grayscale([double value = 0.0]) {
+    conf['grayscale'] = value;
+    return this;
+  }
+  sepia([double value = 0.0]) {
+    conf['sepia'] = value;
+    return this;
+  }
+  hueRotate([double value = 0.0]) {
+    conf['hueRotate'] = value;
+    return this;
+  }
+  brightness([double value = 1.0]) {
+    conf['brightness'] = value;
+    return this;
+  }
+  saturate([double value = 1.0]) {
+    conf['saturate'] = value;
+    return this;
+  }
+  invert([double value = 0.0]) {
+    conf['invert'] = value;
+    return this;
+  }
+  blur([double value = 0.0]) {
+    conf['blur'] = value;
+    return this;
+  }
+  opacity([double value = 1.0]) {
+    conf['opacity'] = value;
+    return this;
+  }
 }