+++ CHANGELOG.md
@@ -18,3 +18,10 @@
 * update test.
 * update example.
 * add CI
+
+## 1.0.0
+
+* add CSS Instagram filters.
+* update docs.
+* support for custom preset.
+* optimize the code structure.